move Dependency and children to separate package
