+++ annotated/src/main/java/com/lezhnin/yadi/annotated/ServiceFromMethodFinder.java
@@ -2,7 +2,7 @@ package com.lezhnin.yadi.annotated;
 
 import static com.lezhnin.yadi.annotated.NamedClassReference.namedClassReference;
 import static com.lezhnin.yadi.annotated.NamedMethodReference.namedMethodReference;
-import static com.lezhnin.yadi.api.MethodDependency.methodFromClass;
+import static com.lezhnin.yadi.api.dependency.MethodDependency.methodFromClass;
 import static com.lezhnin.yadi.api.ServiceDefinition.serviceDefinition;
 import static java.util.Arrays.stream;
 import com.lezhnin.yadi.api.ServiceDefinition;
