+++ api/src/main/java/com/lezhnin/yadi/api/dependency/InstanceMethodDependency.java
@@ -1,8 +1,9 @@
-package com.lezhnin.yadi.api;
+package com.lezhnin.yadi.api.dependency;
 
 import static com.lezhnin.yadi.api.ServiceReference.serviceReference;
 import static com.lezhnin.yadi.api.ServiceReference.toTypes;
 import static java.util.Objects.requireNonNull;
+import com.lezhnin.yadi.api.ServiceReference;
 import java.lang.reflect.Method;
 import javax.annotation.Nonnull;
