+++ simple/src/main/java/com/lezhnin/yadi/simple/SimplePostConstruct.java
@@ -2,9 +2,9 @@ package com.lezhnin.yadi.simple;
 
 import static java.util.Arrays.stream;
 import static java.util.Objects.requireNonNull;
-import com.lezhnin.yadi.api.Dependency;
-import com.lezhnin.yadi.api.InstanceMethodDependency;
-import com.lezhnin.yadi.api.MethodDependency;
+import com.lezhnin.yadi.api.dependency.Dependency;
+import com.lezhnin.yadi.api.dependency.InstanceMethodDependency;
+import com.lezhnin.yadi.api.dependency.MethodDependency;
 import com.lezhnin.yadi.api.MethodNotFoundException;
 import com.lezhnin.yadi.api.ServiceLocator;
 import com.lezhnin.yadi.api.ServiceReference;
